some debugging and changed use to 'real(kind=16)' instead of 'double precision' which caused problems with some compilers

+++ RETRIEVAL/Forward.F90
@@ -36,48 +36,48 @@ program profiling
     use sorting,             only: Sort
     implicit none
     ! variables for setting up vlidort
-    double precision,dimension(:,:,:),allocatable :: greekmat_total_input
-    double precision :: l_omega_total_input (max_atmoswfs, maxlayers) = 0.0
-    double precision :: l_deltau_vert_input (max_atmoswfs, maxlayers) = 0.0
-    double precision :: l_greekmat_total_input(&
+    real(kind=16),dimension(:,:,:),allocatable :: greekmat_total_input
+    real(kind=16) :: l_omega_total_input (max_atmoswfs, maxlayers) = 0.0
+    real(kind=16) :: l_deltau_vert_input (max_atmoswfs, maxlayers) = 0.0
+    real(kind=16) :: l_greekmat_total_input(&
         max_atmoswfs, 0:maxmoments_input, maxlayers, maxstokes_sq) = 0.0
     integer          :: layer_vary_number(maxlayers)
-    double precision :: omega_total_input(maxlayers) = 0.0
+    real(kind=16) :: omega_total_input(maxlayers) = 0.0
     logical          :: layer_vary_flag( maxlayers )
     logical :: openfileflag
     integer :: isgood = 1
     type(vlidort_modified_inputs)          :: vlidort_modin
     type(vlidort_outputs)                  :: vlidort_out
     type(vlidort_linoutputs)               :: vlidort_linout
-    double precision,dimension(:), allocatable    :: deltau_vert_input
+    real(kind=16),dimension(:), allocatable    :: deltau_vert_input
     ! measurement variables
-    double precision,allocatable,dimension(:)   :: SCDerrmeasured
-    double precision,allocatable,dimension(:)   :: SCDmeasured
-    double precision,allocatable,dimension(:)   :: elevangle
-    double precision,allocatable,dimension(:)   :: szangle
-    double precision,allocatable,dimension(:)   :: rel_azimang
+    real(kind=16),allocatable,dimension(:)   :: SCDerrmeasured
+    real(kind=16),allocatable,dimension(:)   :: SCDmeasured
+    real(kind=16),allocatable,dimension(:)   :: elevangle
+    real(kind=16),allocatable,dimension(:)   :: szangle
+    real(kind=16),allocatable,dimension(:)   :: rel_azimang
     !variables for the inversion step
-    double precision :: I90 = 0.0, I090 = 0.0
-    double precision,allocatable,dimension(:)     :: I0alpha
-    double precision,allocatable,dimension(:)     :: Ialpha
-    double precision, allocatable, dimension(:)   :: middleheights, deltaz
-    double precision,dimension(:), allocatable    :: temp
-    double precision,dimension(:), allocatable    :: dbetadC, beta
-    double precision,dimension(:), allocatable    :: temp2
-    double precision,dimension(:), allocatable    :: SCD_gas, SCD_O4
-    double precision,allocatable,dimension(:)     :: abs_coef, abs_coef2, abs_gas, abs_O4
-    double precision,allocatable,dimension(:)     :: Cq_gas, Cq_O4
+    real(kind=16) :: I90 = 0.0, I090 = 0.0
+    real(kind=16),allocatable,dimension(:)     :: I0alpha
+    real(kind=16),allocatable,dimension(:)     :: Ialpha
+    real(kind=16), allocatable, dimension(:)   :: middleheights, deltaz
+    real(kind=16),dimension(:), allocatable    :: temp
+    real(kind=16),dimension(:), allocatable    :: dbetadC, beta
+    real(kind=16),dimension(:), allocatable    :: temp2
+    real(kind=16),dimension(:), allocatable    :: SCD_gas, SCD_O4
+    real(kind=16),allocatable,dimension(:)     :: abs_coef, abs_coef2, abs_gas, abs_O4
+    real(kind=16),allocatable,dimension(:)     :: Cq_gas, Cq_O4
     integer, allocatable, dimension(:) :: skippedrows
-    double precision :: omega
-    double precision :: h1, h2, fact
-    double precision :: tau, temps, temps2
-    double precision :: dtaudC
-    double precision :: cseco4
+    real(kind=16) :: omega
+    real(kind=16) :: h1, h2, fact
+    real(kind=16) :: tau, temps, temps2
+    real(kind=16) :: dtaudC
+    real(kind=16) :: cseco4
     character(len=100) :: cseco4_as_str
-    double precision :: csectg
+    real(kind=16) :: csectg
     character(len=100) :: csectg_as_str
-    double precision :: domegadC
-    double precision :: start, finish, convcrit_scd
+    real(kind=16) :: domegadC
+    real(kind=16) :: start, finish, convcrit_scd
     integer :: num_args, ix, ngr_moms=80, ii, nangle, jj
     integer :: n_user_vzangles, nstart
     character(len=500), dimension(:), allocatable :: args
